Rescale all columns of a matrix

Rescale all columns of a matrix

clamp(.data)

clamp_robust(.data)

clamp_sd(.data, sd = 1e-04)

Arguments

.data
A numeric matrix

sd
the value of each columns standard devation

Details

These functions are used internally by the tour to rescale all columns of .data.

clamp() rescales so all values for each column lie in the unit interval

clamp_robust() rescales by first centering by the median and then scaling by MADs

clamp_sd() rescales all columns to have a fixed standard deviation.

Examples

mv <- matrix(rnorm(300), ncol = 3)

clamp(mv)

#>              [,1]       [,2]       [,3]
#>   [1,] 0.22584703 0.33886386 0.39588962
#>   [2,] 0.53423695 0.24966645 0.74714113
#>   [3,] 0.03261528 0.18889680 0.46624872
#>   [4,] 0.48563404 0.24740218 0.42674488
#>   [5,] 0.60246582 0.03220661 0.00000000
#>   [6,] 0.70061843 0.27092200 0.49300587
#>   [7,] 0.14727146 0.30048787 0.26457492
#>   [8,] 0.44059582 0.30538849 0.35589480
#>   [9,] 0.44117813 0.47647052 0.07516773
#>  [10,] 0.43400457 0.64476103 1.00000000
#>  [11,] 0.38351902 0.37880834 0.25380300
#>  [12,] 0.60384989 0.11214107 0.41330504
#>  [13,] 0.87138141 0.48350611 0.63447581
#>  [14,] 0.18282233 0.32663712 0.23081956
#>  [15,] 0.58213594 0.55295727 0.58654810
#>  [16,] 0.13959714 0.90199106 0.49489363
#>  [17,] 0.38942221 0.52076052 0.58976123
#>  [18,] 0.47687234 0.04833258 0.56167667
#>  [19,] 0.58783094 0.48145984 0.19514442
#>  [20,] 0.31638189 0.66590203 0.66486344
#>  [21,] 0.57388806 0.88239608 0.41363906
#>  [22,] 0.55428895 0.52626965 0.46532306
#>  [23,] 0.24363846 0.73734967 0.48804430
#>  [24,] 0.62411801 0.52114312 0.50454804
#>  [25,] 0.83849614 0.44969396 0.28421463
#>  [26,] 0.46851816 0.43058995 0.72220374
#>  [27,] 0.31232572 0.56079945 0.54696168
#>  [28,] 0.48370045 0.36676482 0.38738772
#>  [29,] 0.33264305 0.30726984 0.61924516
#>  [30,] 0.20491534 0.46863584 0.18010031
#>  [31,] 0.66092800 0.16885796 0.52674838
#>  [32,] 0.51955139 0.61513371 0.22251807
#>  [33,] 0.53207001 0.44390058 0.62675663
#>  [34,] 0.78913543 0.13326194 0.17351431
#>  [35,] 0.50754440 0.21394453 0.49409833
#>  [36,] 0.46170862 0.52736447 0.35464734
#>  [37,] 0.13082700 0.23403865 0.31428693
#>  [38,] 0.43465069 0.36133425 0.61266923
#>  [39,] 0.42827768 0.62605930 0.70650116
#>  [40,] 0.68550899 0.44067171 0.14918233
#>  [41,] 0.49971929 0.38915813 0.30297932
#>  [42,] 0.36760473 0.24028627 0.73873633
#>  [43,] 0.47736361 0.84591612 0.69606354
#>  [44,] 0.43982116 0.59727169 0.71639046
#>  [45,] 0.56953665 0.64509742 0.19927130
#>  [46,] 1.00000000 0.72166291 0.75154679
#>  [47,] 0.49534065 0.17479122 0.48076776
#>  [48,] 0.59429785 0.54086176 0.46486922
#>  [49,] 0.50869140 0.08757166 0.56678339
#>  [50,] 0.13052277 0.76891612 0.50545507
#>  [51,] 0.64727671 0.52172120 0.31722701
#>  [52,] 0.44135751 0.45294618 0.38132556
#>  [53,] 0.44827838 0.17069183 0.24590546
#>  [54,] 0.49024470 0.30060588 0.33263064
#>  [55,] 0.49217906 0.66135508 0.76585921
#>  [56,] 0.58910358 0.57739103 0.31754311
#>  [57,] 0.06300952 0.43421968 0.53702175
#>  [58,] 0.98642627 0.34577003 0.38594511
#>  [59,] 0.41937726 0.60094601 0.36145818
#>  [60,] 0.52641965 0.60577693 0.76669695
#>  [61,] 0.68682016 0.00000000 0.50952846
#>  [62,] 0.36276753 0.63520320 0.19886893
#>  [63,] 0.69419877 0.35637144 0.47813230
#>  [64,] 0.44086201 0.48452065 0.43850355
#>  [65,] 0.26729459 0.02452004 0.30228750
#>  [66,] 0.30487320 0.24903341 0.06761259
#>  [67,] 0.68508972 0.17180971 0.85172114
#>  [68,] 0.51120190 0.50705096 0.29016224
#>  [69,] 0.57770240 0.36155987 0.46001873
#>  [70,] 0.17006817 0.54154926 0.52724571
#>  [71,] 0.21267712 0.40254554 0.49210554
#>  [72,] 0.53960852 0.20741351 0.05281093
#>  [73,] 0.73581172 0.71021504 0.45702457
#>  [74,] 0.53076794 0.64082100 0.50219184
#>  [75,] 0.73226702 0.59778265 0.48170194
#>  [76,] 0.58427516 0.65763509 0.84865642
#>  [77,] 0.59970774 0.31944979 0.90380440
#>  [78,] 0.46619119 0.13993286 0.61943384
#>  [79,] 0.51874902 0.35713992 0.63067270
#>  [80,] 0.46984419 0.92101214 0.41966865
#>  [81,] 0.84517276 0.19227201 0.67875691
#>  [82,] 0.72855958 0.16878535 0.29979292
#>  [83,] 0.62617007 0.05033824 0.59071610
#>  [84,] 0.59029529 0.76739235 0.54344066
#>  [85,] 0.38453287 0.54972859 0.55211314
#>  [86,] 0.69356209 0.87203449 0.70188137
#>  [87,] 0.00000000 0.39107451 0.29538786
#>  [88,] 0.45766656 1.00000000 0.52290434
#>  [89,] 0.56750464 0.66347417 0.75792511
#>  [90,] 0.41551533 0.68154927 0.40484773
#>  [91,] 0.56569714 0.67732569 0.44599032
#>  [92,] 0.68472545 0.33667146 0.43663754
#>  [93,] 0.68204474 0.24119322 0.28203703
#>  [94,] 0.47957348 0.29593131 0.57449532
#>  [95,] 0.57724040 0.82066872 0.79501487
#>  [96,] 0.79832619 0.45538515 0.18283486
#>  [97,] 0.42065528 0.50992852 0.20226346
#>  [98,] 0.66297442 0.68065617 0.65221946
#>  [99,] 0.73199372 0.70362764 0.60263859
#> [100,] 0.43140859 0.56269887 0.89468297

clamp_robust(mv)

#>               [,1]         [,2]         [,3]
#>   [1,] -2.00665511 -0.467281192 -0.391427950
#>   [2,]  0.22108437 -0.828769645  1.219554477
#>   [3,] -3.40251786 -1.075049506 -0.068732489
#>   [4,] -0.13001210 -0.837946014 -0.249913193
#>   [5,]  0.71395443 -1.710064564 -2.207138957
#>   [6,]  1.42298686 -0.742627716  0.053986709
#>   [7,] -2.57426734 -0.622806751 -0.993690672
#>   [8,] -0.45535812 -0.602946111 -0.574860480
#>   [9,] -0.45115164  0.090394428 -1.862389153
#>  [10,] -0.50297176  0.772421818  2.379268205
#>  [11,] -0.86766807 -0.305399019 -1.043095046
#>  [12,]  0.72395260 -1.386115834 -0.311553755
#>  [13,]  2.65654023  0.118907397  0.702825458
#>  [14,] -2.31745583 -0.516832219 -1.148506487
#>  [15,]  0.56709591  0.400370656  0.483009457
#>  [16,] -2.62970489  1.814892527  0.062644715
#>  [17,] -0.82502474  0.269887548  0.497746186
#>  [18,] -0.19330464 -1.644711186  0.368938959
#>  [19,]  0.60823538  0.110614527 -1.312127191
#>  [20,] -1.35265159  0.858099492  0.842195476
#>  [21,]  0.50751514  1.735480355 -0.310021832
#>  [22,]  0.36593557  0.292214278 -0.072977954
#>  [23,] -1.87813383  1.147653840  0.031230917
#>  [24,]  0.87036500  0.271438102  0.106923804
#>  [25,]  2.41898444 -0.018122384 -0.903614958
#>  [26,] -0.25365340 -0.095544797  1.105181456
#>  [27,] -1.38195247  0.432152474  0.301449996
#>  [28,] -0.14397991 -0.354207536 -0.430421135
#>  [29,] -1.23518462 -0.595321590  0.632971467
#>  [30,] -2.15786096  0.058642965 -1.381125603
#>  [31,]  1.13627207 -1.156260506  0.208743565
#>  [32,]  0.11499918  0.652351805 -1.186580467
#>  [33,]  0.20543089 -0.041601089  0.667422117
#>  [34,]  2.06241376 -1.300519751 -1.411331697
#>  [35,]  0.02826341 -0.973539110  0.058997145
#>  [36,] -0.30284397  0.296651256 -0.580581860
#>  [37,] -2.69305846 -0.892104113 -0.765691112
#>  [38,] -0.49830434 -0.376215923  0.602811597
#>  [39,] -0.54434153  0.696629714  1.033163017
#>  [40,]  1.31383966 -0.054686662 -1.522928044
#>  [41,] -0.02826341 -0.263454647 -0.817552431
#>  [42,] -0.98262930 -0.866784493  1.181006616
#>  [43,] -0.18975582  1.587638805  0.985291854
#>  [44,] -0.46095403  0.579962748  1.078519382
#>  [45,]  0.47608151  0.773785074 -1.293199628
#>  [46,]  3.58565203  1.084080442  1.239760627
#>  [47,] -0.05989374 -1.132214924 -0.002142282
#>  [48,]  0.65495082  0.351351428 -0.075059435
#>  [49,]  0.03654905 -1.485687765  0.392360423
#>  [50,] -2.69525614  1.275582518  0.111083774
#>  [51,]  1.03765825  0.273780878 -0.752206734
#>  [52,] -0.44985584 -0.004942193 -0.458224680
#>  [53,] -0.39986099 -1.148828409 -1.079316376
#>  [54,] -0.09670573 -0.622328483 -0.681559402
#>  [55,] -0.08273235  0.839672167  1.305403215
#>  [56,]  0.61742862  0.499392853 -0.750756967
#>  [57,] -3.18295667 -0.080834671  0.255861430
#>  [58,]  3.48759848 -0.439292698 -0.437037551
#>  [59,] -0.60863621  0.594853601 -0.549344559
#>  [60,]  0.16461398  0.614431787  1.309245413
#>  [61,]  1.32331121 -1.840587592  0.129766038
#>  [62,] -1.01757219  0.733687005 -1.295045062
#>  [63,]  1.37661262 -0.396328573 -0.014229554
#>  [64,] -0.45343515  0.123019008 -0.195983119
#>  [65,] -1.70724726 -1.741215740 -0.820725384
#>  [66,] -1.43578778 -0.831335154 -1.897040071
#>  [67,]  1.31081092 -1.144298009  1.699200963
#>  [68,]  0.05468436  0.214327135 -0.876336797
#>  [69,]  0.53506902 -0.375301529 -0.097305754
#>  [70,] -2.40958902  0.354137639  0.211024547
#>  [71,] -2.10179155 -0.209199785  0.049857415
#>  [72,]  0.25988743 -1.000007233 -1.964926539
#>  [73,]  1.67721522  1.037685875 -0.111038217
#>  [74,]  0.19602504  0.756454150  0.096117314
#>  [75,]  1.65160916  0.582033510  0.002142282
#>  [76,]  0.58254921  0.824596255  1.685144938
#>  [77,]  0.69403069 -0.545960167  1.938076002
#>  [78,] -0.27046287 -1.273484638  0.633836850
#>  [79,]  0.10920308 -0.393214194  0.685382825
#>  [80,] -0.24407443  1.891978859 -0.282367678
#>  [81,]  2.46721480 -1.061370895  0.905916581
#>  [82,]  1.62482744 -1.156554786 -0.832166549
#>  [83,]  0.88518862 -1.636582871  0.502125617
#>  [84,]  0.62603724  1.269407167  0.285301177
#>  [85,] -0.86034421  0.387285849  0.325076719
#>  [86,]  1.37201340  1.693488159  1.011974794
#>  [87,] -3.63812330 -0.255688194 -0.852369950
#>  [88,] -0.33204289  2.212091277  0.191113264
#>  [89,]  0.46140273  0.848260154  1.269014189
#>  [90,] -0.63653389  0.921512736 -0.350342423
#>  [91,]  0.44834575  0.904395930 -0.161645756
#>  [92,]  1.30817950 -0.476166299 -0.204541418
#>  [93,]  1.28881466 -0.863108944 -0.913602289
#>  [94,] -0.17379226 -0.641273033  0.427730507
#>  [95,]  0.53173165  1.485319196  1.439122936
#>  [96,]  2.12880575  0.004942193 -1.368583861
#>  [97,] -0.59940408  0.225988946 -1.279476385
#>  [98,]  1.15105497  0.917893282  0.784205055
#>  [99,]  1.64963488  1.010989275  0.556806967
#> [100,] -0.52172454  0.439850243  1.896241413

clamp_sd(mv)

#>                 [,1]          [,2]          [,3]
#>   [1,] -1.332226e-04 -3.817624e-05 -4.085038e-05
#>   [2,]  2.429497e-05 -7.743754e-05  1.294316e-04
#>   [3,] -2.319204e-04 -1.041860e-04 -6.741237e-06
#>   [4,] -5.301417e-07 -7.843419e-05 -2.589217e-05
#>   [5,]  5.914451e-05 -1.731551e-04 -2.327723e-04
#>   [6,]  1.092783e-04 -6.808166e-05  6.230270e-06
#>   [7,] -1.733570e-04 -5.506789e-05 -1.045100e-04
#>   [8,] -2.353450e-05 -5.291082e-05 -6.023933e-05
#>   [9,] -2.323707e-05  2.239297e-05 -1.963320e-04
#>  [10,] -2.690114e-05  9.646804e-05  2.520142e-04
#>  [11,] -5.268786e-05 -2.059420e-05 -1.097321e-04
#>  [12,]  5.985146e-05 -1.379710e-04 -3.240762e-05
#>  [13,]  1.964996e-04  2.548976e-05  7.481299e-05
#>  [14,] -1.551985e-04 -4.355799e-05 -1.208741e-04
#>  [15,]  4.876053e-05  5.605952e-05  5.157828e-05
#>  [16,] -1.772768e-04  2.096909e-04  7.145428e-06
#>  [17,] -4.967266e-05  4.188773e-05  5.313597e-05
#>  [18,] -5.005391e-06 -1.660570e-04  3.952095e-05
#>  [19,]  5.166940e-05  2.458907e-05 -1.381689e-04
#>  [20,] -8.697976e-05  1.057735e-04  8.954450e-05
#>  [21,]  4.454773e-05  2.010659e-04 -3.224569e-05
#>  [22,]  3.453701e-05  4.431264e-05 -7.189985e-06
#>  [23,] -1.241352e-04  1.372220e-04  3.824967e-06
#>  [24,]  7.020389e-05  4.205614e-05  1.182576e-05
#>  [25,]  1.797027e-04  1.060694e-05 -9.498891e-05
#>  [26,] -9.272493e-06  2.198087e-06  1.173423e-04
#>  [27,] -8.905155e-05  5.951135e-05  3.238732e-05
#>  [28,] -1.517769e-06 -2.589530e-05 -4.497199e-05
#>  [29,] -7.867398e-05 -5.208272e-05  6.742937e-05
#>  [30,] -1.439140e-04  1.894444e-05 -1.454621e-04
#>  [31,]  8.900548e-05 -1.130064e-04  2.258818e-05
#>  [32,]  1.679396e-05  8.342722e-05 -1.248986e-04
#>  [33,]  2.318815e-05  8.056919e-06  7.107083e-05
#>  [34,]  1.544905e-04 -1.286744e-04 -1.486549e-04
#>  [35,]  1.066110e-05 -9.316097e-05  6.759877e-06
#>  [36,] -1.275063e-05  4.479454e-05 -6.084408e-05
#>  [37,] -1.817564e-04 -8.431631e-05 -8.041026e-05
#>  [38,] -2.657112e-05 -2.828563e-05  6.424145e-05
#>  [39,] -2.982629e-05  8.823625e-05  1.097299e-04
#>  [40,]  1.015608e-04  6.635694e-06 -1.604507e-04
#>  [41,]  6.664241e-06 -1.603862e-05 -8.589204e-05
#>  [42,] -6.081646e-05 -8.156634e-05  1.253571e-04
#>  [43,] -4.754463e-06  1.850088e-04  1.046699e-04
#>  [44,] -2.393018e-05  7.556504e-05  1.145241e-04
#>  [45,]  4.232514e-05  9.661610e-05 -1.361683e-04
#>  [46,]  2.621947e-04  1.303173e-04  1.315674e-04
#>  [47,]  4.427743e-06 -1.103948e-04  2.973958e-07
#>  [48,]  5.497252e-05  5.073553e-05 -7.409999e-06
#>  [49,]  1.124696e-05 -1.487855e-04  4.199661e-05
#>  [50,] -1.819118e-04  1.511164e-04  1.226547e-05
#>  [51,]  8.203276e-05  4.231059e-05 -7.898495e-05
#>  [52,] -2.314545e-05  1.203845e-05 -4.791084e-05
#>  [53,] -1.961045e-05 -1.121992e-04 -1.135607e-04
#>  [54,]  1.824864e-06 -5.501595e-05 -7.151748e-05
#>  [55,]  2.812885e-06  1.037721e-04  1.385059e-04
#>  [56,]  5.231943e-05  6.681433e-05 -7.883171e-05
#>  [57,] -2.163958e-04  3.795756e-06  2.756857e-05
#>  [58,]  2.552616e-04 -3.513641e-05 -4.567135e-05
#>  [59,] -3.437239e-05  7.718233e-05 -5.754228e-05
#>  [60,]  2.030209e-05  7.930872e-05  1.389120e-04
#>  [61,]  1.022305e-04 -1.873312e-04  1.424020e-05
#>  [62,] -6.328718e-05  9.226104e-05 -1.363634e-04
#>  [63,]  1.059993e-04 -3.047007e-05 -9.802375e-07
#>  [64,] -2.339854e-05  2.593633e-05 -2.019172e-05
#>  [65,] -1.120523e-04 -1.765384e-04 -8.622742e-05
#>  [66,] -9.285811e-05 -7.771618e-05 -1.999947e-04
#>  [67,]  1.013467e-04 -1.117071e-04  1.801306e-04
#>  [68,]  1.252926e-05  3.585331e-05 -9.210559e-05
#>  [69,]  4.649600e-05 -2.818632e-05 -9.761451e-06
#>  [70,] -1.617130e-04  5.103815e-05  2.282928e-05
#>  [71,] -1.399495e-04 -1.014599e-05  5.793801e-06
#>  [72,]  2.703863e-05 -9.603568e-05 -2.071703e-04
#>  [73,]  1.272541e-04  1.252784e-04 -1.121298e-05
#>  [74,]  2.252309e-05  9.473379e-05  1.068351e-05
#>  [75,]  1.254436e-04  7.578994e-05  7.502774e-07
#>  [76,]  4.985319e-05  1.021347e-04  1.786449e-04
#>  [77,]  5.773576e-05 -4.672158e-05  2.053799e-04
#>  [78,] -1.046105e-05 -1.257381e-04  6.752085e-05
#>  [79,]  1.638413e-05 -3.013182e-05  7.296929e-05
#>  [80,] -8.595189e-06  2.180632e-04 -2.932263e-05
#>  [81,]  1.831129e-04 -1.027004e-04  9.627987e-05
#>  [82,]  1.235499e-04 -1.130383e-04 -8.743676e-05
#>  [83,]  7.125203e-05 -1.651742e-04  5.359887e-05
#>  [84,]  5.292812e-05  1.504457e-04  3.068038e-05
#>  [85,] -5.217000e-05  5.463838e-05  3.488468e-05
#>  [86,]  1.056741e-04  1.965052e-04  1.074903e-04
#>  [87,] -2.485795e-04 -1.519510e-05 -8.957228e-05
#>  [88,] -1.481521e-05  2.528307e-04  2.072465e-05
#>  [89,]  4.128725e-05  1.047048e-04  1.346595e-04
#>  [90,] -3.634497e-05  1.126608e-04 -3.650761e-05
#>  [91,]  4.036402e-05  1.108018e-04 -1.656224e-05
#>  [92,]  1.011606e-04 -3.914126e-05 -2.109634e-05
#>  [93,]  9.979137e-05 -8.116714e-05 -9.604458e-05
#>  [94,] -3.625722e-06 -5.707352e-05  4.573526e-05
#>  [95,]  4.626002e-05  1.738959e-04  1.526402e-04
#>  [96,]  1.591849e-04  1.311199e-05 -1.441365e-04
#>  [97,] -3.371961e-05  3.711990e-05 -1.347177e-04
#>  [98,]  9.005074e-05  1.122677e-04  8.341487e-05
#>  [99,]  1.253040e-04  1.223789e-04  5.937873e-05
#> [100,] -2.822710e-05  6.034740e-05  2.009579e-04