Tour a high dimensional dataset

Tour a high dimensional dataset

limn_tour(
  .data,
  cols,
  color = NULL,
  tour_path = tourr::grand_tour(),
  rescale = clamp,
  morph = "center"
)

Arguments

.data
a data.frame to tour

cols
Columns to tour. This can use a tidyselect specification such as tidyselect::starts_with().

color
A variable name in .data, mapping to the color aesthetic, if NULL points will be colored black.

tour_path
the tour path to take, the default is tourr::grand_tour() but also works with tourr::guided_tour().

rescale
A function that rescales cols, the default is to clamp() the data to lie in the hyperdimensional unit cube. To not perform any scaling use identity().

morph
One of c("center", "centre", "identity", "radial") that rescales each projection along the tour path. The default is to center the projections and divide by half range. See morph_center() for details.

Value

A shiny app object

Details

The tour interface consists of two views

the tour view which is a dynamic scatterplot

the axis view which shows the direction and magnitude of the basis vectors being generated. There are several other user controls available:

There is a play button, that when pressed will start the tour.

There is also a text view of the half range which is the maximum squared Euclidean distance between points in the tour view. The half range is a scale factor for projections and can be thought of as a way of zooming in and out on points. It can be dynamically modified by scrolling (via a mouse-wheel). To reset double click the tour view.

The legend can be toggled to highlight groups of points with shift+mouse-click. Multiple groups can be selected in this way. To reset double click the legend title.

See also

Examples

if (FALSE) {
# tour the first ten columns of the fake tree data
# loads the default interface
limn_tour(fake_trees, dim1:dim10)
# perform the same action but now coloring points
limn_tour(fake_trees, dim1:dim10, color = branches)
}