Link a 2-d embedding with a tour

Link a 2-d embedding with a tour

limn_tour_link(
  embed_data,
  tour_data,
  cols,
  color = NULL,
  tour_path = tourr::grand_tour(),
  rescale = clamp,
  morph = "center",
  reference = NULL
)

Arguments

embed_data
A data.frame representing embedding coordinates

tour_data
A data.frame that is linked to embed_data to tour

cols
Columns in tour_data to tour. This can use a tidyselect specification such as tidyselect::starts_with().

color
An optional bare column name in either embed_data or tour_data for color mapping the points in the views. If NULL points will be colored black.

tour_path
the tour path to take, the default is tourr::grand_tour().

rescale
A function that rescales tour columns. Default is clamp() To not perform any scaling use identity().

morph
One of c("center", "centre", "identity", "radial") that rescales each projection along the tour path. The default is to center the projections and divide by half range. See morph_center() for details for each of these functions.

reference
an optional data set to compute a nearest neighbors from. By default the reference is set to NULL in which case neighbors will be computed from tour_data. Otherwise data.frame, matrix or dist objects can be used.

Value

After pressing the Done button on the interface, a list of artefacts is returned to the R session.

Details

The tour view on the left is a dynamic and interactive scatterplot. Brushing on the tour view is activated with the shift key plus a mouse drag. By default it will highlight corresponding points in the xy view and pause the animation.

The xy view on the right is an interactive scatterplot. Brushing on the xy view will highlight points in the tour view and is activated via a mouse drag, the type of highlighting depends on the brush mode selected.

the axis view which shows the direction and magnitude of the basis vectors being generated. There are several other user controls available:

There is a play button, that when pressed will start the tour.

There is also a text view of the half range which is the maximum squared Euclidean distance between points in the tour view. The half range is a scale factor for projections and can be thought of as a way of zooming in and out on points. It can be dynamically modified by scrolling (via a mouse-wheel). To reset double click the tour view.

The legend can be toggled to highlight groups of points with shift+mouse-click. Multiple groups can be selected in this way. To reset double click the legend title.

There are different brushing modes that can be modified via radio buttons. Both neighbors and distance based brushing operate on the k-NN graph obtained from the control panel.

Examples

if (FALSE) {
# tour the first ten columns of the fake tree data and link to the
# another layout based on t-SNE
# loads the default interface
tsne <- Rtsne::Rtsne(dplyr::select(fake_trees, dplyr::starts_with("dim")))
tsne_df <- data.frame(tsneX = tsne$Y[,1], tsneY = tsne$Y[,2])
limn_tour_link(
  tsne_df,
  dplyr::select(fake_trees, dim1:dim10, branches),
  color = branches,
)
# assigning to an object will return a list of artefacts after clicking
# done on the tourr
res <- limn_tour_link(
  tsne_df,
  dplyr::select(fake_trees, dim1:dim10, branches),
  color = branches,
)

}