Rescale all columns of a matrix

Rescale all columns of a matrix

clamp(.data)

clamp_robust(.data)

clamp_sd(.data, sd = 1)

clamp_standardize(.data, sd = 1)

Arguments

.data
A numeric matrix

sd
the value of each columns standard deviation (default is 1)

Details

These functions are used internally by the tour to rescale all columns of .data.

clamp() rescales so all values for each column lie in the unit interval

clamp_robust() rescales by first centering by the median and then scaling by the median absolute deviation.

clamp_sd() rescales all columns to have a fixed standard deviation.

clamp_standardize() rescales all columns to have zero mean and unit variance.

Examples

mv <- matrix(rnorm(300), ncol = 3)

clamp(mv)
#>              [,1]       [,2]       [,3]
#>   [1,] 0.22584703 0.33886386 0.39588962
#>   [2,] 0.53423695 0.24966645 0.74714113
#>   [3,] 0.03261528 0.18889680 0.46624872
#>   [4,] 0.48563404 0.24740218 0.42674488
#>   [5,] 0.60246582 0.03220661 0.00000000
#>   [6,] 0.70061843 0.27092200 0.49300587
#>   [7,] 0.14727146 0.30048787 0.26457492
#>   [8,] 0.44059582 0.30538849 0.35589480
#>   [9,] 0.44117813 0.47647052 0.07516773
#>  [10,] 0.43400457 0.64476103 1.00000000
#>  [11,] 0.38351902 0.37880834 0.25380300
#>  [12,] 0.60384989 0.11214107 0.41330504
#>  [13,] 0.87138141 0.48350611 0.63447581
#>  [14,] 0.18282233 0.32663712 0.23081956
#>  [15,] 0.58213594 0.55295727 0.58654810
#>  [16,] 0.13959714 0.90199106 0.49489363
#>  [17,] 0.38942221 0.52076052 0.58976123
#>  [18,] 0.47687234 0.04833258 0.56167667
#>  [19,] 0.58783094 0.48145984 0.19514442
#>  [20,] 0.31638189 0.66590203 0.66486344
#>  [21,] 0.57388806 0.88239608 0.41363906
#>  [22,] 0.55428895 0.52626965 0.46532306
#>  [23,] 0.24363846 0.73734967 0.48804430
#>  [24,] 0.62411801 0.52114312 0.50454804
#>  [25,] 0.83849614 0.44969396 0.28421463
#>  [26,] 0.46851816 0.43058995 0.72220374
#>  [27,] 0.31232572 0.56079945 0.54696168
#>  [28,] 0.48370045 0.36676482 0.38738772
#>  [29,] 0.33264305 0.30726984 0.61924516
#>  [30,] 0.20491534 0.46863584 0.18010031
#>  [31,] 0.66092800 0.16885796 0.52674838
#>  [32,] 0.51955139 0.61513371 0.22251807
#>  [33,] 0.53207001 0.44390058 0.62675663
#>  [34,] 0.78913543 0.13326194 0.17351431
#>  [35,] 0.50754440 0.21394453 0.49409833
#>  [36,] 0.46170862 0.52736447 0.35464734
#>  [37,] 0.13082700 0.23403865 0.31428693
#>  [38,] 0.43465069 0.36133425 0.61266923
#>  [39,] 0.42827768 0.62605930 0.70650116
#>  [40,] 0.68550899 0.44067171 0.14918233
#>  [41,] 0.49971929 0.38915813 0.30297932
#>  [42,] 0.36760473 0.24028627 0.73873633
#>  [43,] 0.47736361 0.84591612 0.69606354
#>  [44,] 0.43982116 0.59727169 0.71639046
#>  [45,] 0.56953665 0.64509742 0.19927130
#>  [46,] 1.00000000 0.72166291 0.75154679
#>  [47,] 0.49534065 0.17479122 0.48076776
#>  [48,] 0.59429785 0.54086176 0.46486922
#>  [49,] 0.50869140 0.08757166 0.56678339
#>  [50,] 0.13052277 0.76891612 0.50545507
#>  [51,] 0.64727671 0.52172120 0.31722701
#>  [52,] 0.44135751 0.45294618 0.38132556
#>  [53,] 0.44827838 0.17069183 0.24590546
#>  [54,] 0.49024470 0.30060588 0.33263064
#>  [55,] 0.49217906 0.66135508 0.76585921
#>  [56,] 0.58910358 0.57739103 0.31754311
#>  [57,] 0.06300952 0.43421968 0.53702175
#>  [58,] 0.98642627 0.34577003 0.38594511
#>  [59,] 0.41937726 0.60094601 0.36145818
#>  [60,] 0.52641965 0.60577693 0.76669695
#>  [61,] 0.68682016 0.00000000 0.50952846
#>  [62,] 0.36276753 0.63520320 0.19886893
#>  [63,] 0.69419877 0.35637144 0.47813230
#>  [64,] 0.44086201 0.48452065 0.43850355
#>  [65,] 0.26729459 0.02452004 0.30228750
#>  [66,] 0.30487320 0.24903341 0.06761259
#>  [67,] 0.68508972 0.17180971 0.85172114
#>  [68,] 0.51120190 0.50705096 0.29016224
#>  [69,] 0.57770240 0.36155987 0.46001873
#>  [70,] 0.17006817 0.54154926 0.52724571
#>  [71,] 0.21267712 0.40254554 0.49210554
#>  [72,] 0.53960852 0.20741351 0.05281093
#>  [73,] 0.73581172 0.71021504 0.45702457
#>  [74,] 0.53076794 0.64082100 0.50219184
#>  [75,] 0.73226702 0.59778265 0.48170194
#>  [76,] 0.58427516 0.65763509 0.84865642
#>  [77,] 0.59970774 0.31944979 0.90380440
#>  [78,] 0.46619119 0.13993286 0.61943384
#>  [79,] 0.51874902 0.35713992 0.63067270
#>  [80,] 0.46984419 0.92101214 0.41966865
#>  [81,] 0.84517276 0.19227201 0.67875691
#>  [82,] 0.72855958 0.16878535 0.29979292
#>  [83,] 0.62617007 0.05033824 0.59071610
#>  [84,] 0.59029529 0.76739235 0.54344066
#>  [85,] 0.38453287 0.54972859 0.55211314
#>  [86,] 0.69356209 0.87203449 0.70188137
#>  [87,] 0.00000000 0.39107451 0.29538786
#>  [88,] 0.45766656 1.00000000 0.52290434
#>  [89,] 0.56750464 0.66347417 0.75792511
#>  [90,] 0.41551533 0.68154927 0.40484773
#>  [91,] 0.56569714 0.67732569 0.44599032
#>  [92,] 0.68472545 0.33667146 0.43663754
#>  [93,] 0.68204474 0.24119322 0.28203703
#>  [94,] 0.47957348 0.29593131 0.57449532
#>  [95,] 0.57724040 0.82066872 0.79501487
#>  [96,] 0.79832619 0.45538515 0.18283486
#>  [97,] 0.42065528 0.50992852 0.20226346
#>  [98,] 0.66297442 0.68065617 0.65221946
#>  [99,] 0.73199372 0.70362764 0.60263859
#> [100,] 0.43140859 0.56269887 0.89468297

clamp_robust(mv)
#>               [,1]         [,2]         [,3]
#>   [1,] -2.00665511 -0.467281192 -0.391427950
#>   [2,]  0.22108437 -0.828769645  1.219554477
#>   [3,] -3.40251786 -1.075049506 -0.068732489
#>   [4,] -0.13001210 -0.837946014 -0.249913193
#>   [5,]  0.71395443 -1.710064564 -2.207138957
#>   [6,]  1.42298686 -0.742627716  0.053986709
#>   [7,] -2.57426734 -0.622806751 -0.993690672
#>   [8,] -0.45535812 -0.602946111 -0.574860480
#>   [9,] -0.45115164  0.090394428 -1.862389153
#>  [10,] -0.50297176  0.772421818  2.379268205
#>  [11,] -0.86766807 -0.305399019 -1.043095046
#>  [12,]  0.72395260 -1.386115834 -0.311553755
#>  [13,]  2.65654023  0.118907397  0.702825458
#>  [14,] -2.31745583 -0.516832219 -1.148506487
#>  [15,]  0.56709591  0.400370656  0.483009457
#>  [16,] -2.62970489  1.814892527  0.062644715
#>  [17,] -0.82502474  0.269887548  0.497746186
#>  [18,] -0.19330464 -1.644711186  0.368938959
#>  [19,]  0.60823538  0.110614527 -1.312127191
#>  [20,] -1.35265159  0.858099492  0.842195476
#>  [21,]  0.50751514  1.735480355 -0.310021832
#>  [22,]  0.36593557  0.292214278 -0.072977954
#>  [23,] -1.87813383  1.147653840  0.031230917
#>  [24,]  0.87036500  0.271438102  0.106923804
#>  [25,]  2.41898444 -0.018122384 -0.903614958
#>  [26,] -0.25365340 -0.095544797  1.105181456
#>  [27,] -1.38195247  0.432152474  0.301449996
#>  [28,] -0.14397991 -0.354207536 -0.430421135
#>  [29,] -1.23518462 -0.595321590  0.632971467
#>  [30,] -2.15786096  0.058642965 -1.381125603
#>  [31,]  1.13627207 -1.156260506  0.208743565
#>  [32,]  0.11499918  0.652351805 -1.186580467
#>  [33,]  0.20543089 -0.041601089  0.667422117
#>  [34,]  2.06241376 -1.300519751 -1.411331697
#>  [35,]  0.02826341 -0.973539110  0.058997145
#>  [36,] -0.30284397  0.296651256 -0.580581860
#>  [37,] -2.69305846 -0.892104113 -0.765691112
#>  [38,] -0.49830434 -0.376215923  0.602811597
#>  [39,] -0.54434153  0.696629714  1.033163017
#>  [40,]  1.31383966 -0.054686662 -1.522928044
#>  [41,] -0.02826341 -0.263454647 -0.817552431
#>  [42,] -0.98262930 -0.866784493  1.181006616
#>  [43,] -0.18975582  1.587638805  0.985291854
#>  [44,] -0.46095403  0.579962748  1.078519382
#>  [45,]  0.47608151  0.773785074 -1.293199628
#>  [46,]  3.58565203  1.084080442  1.239760627
#>  [47,] -0.05989374 -1.132214924 -0.002142282
#>  [48,]  0.65495082  0.351351428 -0.075059435
#>  [49,]  0.03654905 -1.485687765  0.392360423
#>  [50,] -2.69525614  1.275582518  0.111083774
#>  [51,]  1.03765825  0.273780878 -0.752206734
#>  [52,] -0.44985584 -0.004942193 -0.458224680
#>  [53,] -0.39986099 -1.148828409 -1.079316376
#>  [54,] -0.09670573 -0.622328483 -0.681559402
#>  [55,] -0.08273235  0.839672167  1.305403215
#>  [56,]  0.61742862  0.499392853 -0.750756967
#>  [57,] -3.18295667 -0.080834671  0.255861430
#>  [58,]  3.48759848 -0.439292698 -0.437037551
#>  [59,] -0.60863621  0.594853601 -0.549344559
#>  [60,]  0.16461398  0.614431787  1.309245413
#>  [61,]  1.32331121 -1.840587592  0.129766038
#>  [62,] -1.01757219  0.733687005 -1.295045062
#>  [63,]  1.37661262 -0.396328573 -0.014229554
#>  [64,] -0.45343515  0.123019008 -0.195983119
#>  [65,] -1.70724726 -1.741215740 -0.820725384
#>  [66,] -1.43578778 -0.831335154 -1.897040071
#>  [67,]  1.31081092 -1.144298009  1.699200963
#>  [68,]  0.05468436  0.214327135 -0.876336797
#>  [69,]  0.53506902 -0.375301529 -0.097305754
#>  [70,] -2.40958902  0.354137639  0.211024547
#>  [71,] -2.10179155 -0.209199785  0.049857415
#>  [72,]  0.25988743 -1.000007233 -1.964926539
#>  [73,]  1.67721522  1.037685875 -0.111038217
#>  [74,]  0.19602504  0.756454150  0.096117314
#>  [75,]  1.65160916  0.582033510  0.002142282
#>  [76,]  0.58254921  0.824596255  1.685144938
#>  [77,]  0.69403069 -0.545960167  1.938076002
#>  [78,] -0.27046287 -1.273484638  0.633836850
#>  [79,]  0.10920308 -0.393214194  0.685382825
#>  [80,] -0.24407443  1.891978859 -0.282367678
#>  [81,]  2.46721480 -1.061370895  0.905916581
#>  [82,]  1.62482744 -1.156554786 -0.832166549
#>  [83,]  0.88518862 -1.636582871  0.502125617
#>  [84,]  0.62603724  1.269407167  0.285301177
#>  [85,] -0.86034421  0.387285849  0.325076719
#>  [86,]  1.37201340  1.693488159  1.011974794
#>  [87,] -3.63812330 -0.255688194 -0.852369950
#>  [88,] -0.33204289  2.212091277  0.191113264
#>  [89,]  0.46140273  0.848260154  1.269014189
#>  [90,] -0.63653389  0.921512736 -0.350342423
#>  [91,]  0.44834575  0.904395930 -0.161645756
#>  [92,]  1.30817950 -0.476166299 -0.204541418
#>  [93,]  1.28881466 -0.863108944 -0.913602289
#>  [94,] -0.17379226 -0.641273033  0.427730507
#>  [95,]  0.53173165  1.485319196  1.439122936
#>  [96,]  2.12880575  0.004942193 -1.368583861
#>  [97,] -0.59940408  0.225988946 -1.279476385
#>  [98,]  1.15105497  0.917893282  0.784205055
#>  [99,]  1.64963488  1.010989275  0.556806967
#> [100,] -0.52172454  0.439850243  1.896241413

clamp_sd(mv)
#>                [,1]        [,2]         [,3]
#>   [1,] -1.332226336 -0.38176243 -0.408503782
#>   [2,]  0.242949666 -0.77437543  1.294316186
#>   [3,] -2.319204176 -1.04186026 -0.067412368
#>   [4,] -0.005301417 -0.78434190 -0.258921670
#>   [5,]  0.591445119 -1.73155081 -2.327723326
#>   [6,]  1.092783309 -0.68081659  0.062302704
#>   [7,] -1.733570020 -0.55067891 -1.045099794
#>   [8,] -0.235345028 -0.52910825 -0.602393280
#>   [9,] -0.232370740  0.22392967 -1.963320313
#>  [10,] -0.269011384  0.96468036  2.520141905
#>  [11,] -0.526878552 -0.20594203 -1.097320571
#>  [12,]  0.598514569 -1.37970974 -0.324076185
#>  [13,]  1.964996457  0.25489763  0.748129911
#>  [14,] -1.551985356 -0.43557986 -1.208741219
#>  [15,]  0.487605328  0.56059518  0.515782822
#>  [16,] -1.772768449  2.09690893  0.071454278
#>  [17,] -0.496726574  0.41887734  0.531359651
#>  [18,] -0.050053908 -1.66057043  0.395209489
#>  [19,]  0.516693960  0.24589074 -1.381689476
#>  [20,] -0.869797648  1.05773497  0.895445018
#>  [21,]  0.445477330  2.01065928 -0.322456931
#>  [22,]  0.345370137  0.44312642 -0.071899855
#>  [23,] -1.241352320  1.37222026  0.038249669
#>  [24,]  0.702038925  0.42056140  0.118257593
#>  [25,]  1.797027001  0.10606944 -0.949889119
#>  [26,] -0.092724928  0.02198087  1.173423083
#>  [27,] -0.890515528  0.59511345  0.323873171
#>  [28,] -0.015177689 -0.25895301 -0.449719857
#>  [29,] -0.786739844 -0.52082725  0.674293741
#>  [30,] -1.439140014  0.18944437 -1.454621292
#>  [31,]  0.890054816 -1.13006362  0.225881810
#>  [32,]  0.167939619  0.83427218 -1.248985690
#>  [33,]  0.231881502  0.08056919  0.710708326
#>  [34,]  1.544905251 -1.28674374 -1.486549350
#>  [35,]  0.106611032 -0.93160971  0.067598771
#>  [36,] -0.127506287  0.44794543 -0.608440820
#>  [37,] -1.817564096 -0.84316307 -0.804102626
#>  [38,] -0.265711175 -0.28285634  0.642414544
#>  [39,] -0.298262863  0.88236247  1.097299077
#>  [40,]  1.015608192  0.06635694 -1.604507488
#>  [41,]  0.066642409 -0.16038620 -0.858920411
#>  [42,] -0.608164613 -0.81566340  1.253570821
#>  [43,] -0.047544632  1.85008841  1.046698915
#>  [44,] -0.239301752  0.75565035  1.145241080
#>  [45,]  0.423251438  0.96616099 -1.361682906
#>  [46,]  2.621946973  1.30317313  1.315674232
#>  [47,]  0.044277431 -1.10394768  0.002973958
#>  [48,]  0.549725225  0.50735535 -0.074099995
#>  [49,]  0.112469593 -1.48785492  0.419966144
#>  [50,] -1.819118017  1.51116374  0.122654711
#>  [51,]  0.820327584  0.42310589 -0.789849541
#>  [52,] -0.231454513  0.12038446 -0.479108403
#>  [53,] -0.196104468 -1.12199161 -1.135606777
#>  [54,]  0.018248642 -0.55015946 -0.715174805
#>  [55,]  0.028128851  1.03772103  1.385058917
#>  [56,]  0.523194260  0.66814330 -0.788317127
#>  [57,] -2.163958240  0.03795756  0.275685730
#>  [58,]  2.552615891 -0.35136410 -0.456713456
#>  [59,] -0.343723937  0.77182332 -0.575422768
#>  [60,]  0.203020939  0.79308721  1.389120148
#>  [61,]  1.022305279 -1.87331201  0.142401998
#>  [62,] -0.632871814  0.92261043 -1.363633544
#>  [63,]  1.059993302 -0.30470072 -0.009802375
#>  [64,] -0.233985352  0.25936326 -0.201917195
#>  [65,] -1.120522922 -1.76538413 -0.862274245
#>  [66,] -0.928581059 -0.77716184 -1.999946581
#>  [67,]  1.013466649 -1.11707113  1.801305958
#>  [68,]  0.125292596  0.35853311 -0.921055907
#>  [69,]  0.464959953 -0.28186322 -0.097614514
#>  [70,] -1.617130308  0.51038145  0.228292824
#>  [71,] -1.399494814 -0.10145993  0.057938011
#>  [72,]  0.270386289 -0.96035676 -2.071703065
#>  [73,]  1.272541498  1.25278394 -0.112129825
#>  [74,]  0.225230872  0.94733785  0.106835055
#>  [75,]  1.254436126  0.75789941  0.007502774
#>  [76,]  0.498531944  1.02134707  1.786448639
#>  [77,]  0.577357573 -0.46721575  2.053798586
#>  [78,] -0.104610461 -1.25738087  0.675208457
#>  [79,]  0.163841347 -0.30131818  0.729692923
#>  [80,] -0.085951893  2.18063248 -0.293226292
#>  [81,]  1.831129425 -1.02700391  0.962798682
#>  [82,]  1.235499474 -1.13038323 -0.874367639
#>  [83,]  0.712520315 -1.65174226  0.535988740
#>  [84,]  0.529281191  1.50445668  0.306803754
#>  [85,] -0.521700045  0.54638376  0.348846788
#>  [86,]  1.056741318  1.96505151  1.074902974
#>  [87,] -2.485794588 -0.15195105 -0.895722778
#>  [88,] -0.148152073  2.52830692  0.207246455
#>  [89,]  0.412872456  1.04704846  1.346595456
#>  [90,] -0.363449653  1.12660816 -0.365076085
#>  [91,]  0.403640209  1.10801758 -0.165622357
#>  [92,]  1.011606039 -0.39141256 -0.210963378
#>  [93,]  0.997913675 -0.81167138 -0.960445799
#>  [94,] -0.036257216 -0.57073516  0.457352577
#>  [95,]  0.462600188  1.73895897  1.526401621
#>  [96,]  1.591849285  0.13111990 -1.441364581
#>  [97,] -0.337196144  0.37119902 -1.347177341
#>  [98,]  0.900507422  1.12267707  0.834148728
#>  [99,]  1.253040160  1.22378873  0.593787319
#> [100,] -0.282270978  0.60347401  2.009579125

clamp_standardize(mv)